migrate to local MySql db for persistance

+++ src/main/java/com/harshraj9988/payment_simulator/dto/AccountDto.java
@@ -10,5 +10,5 @@ import lombok.NoArgsConstructor;
 public class AccountDto {
     private Long id;
     private String name;
-    private Double balance;
+    private Long balance;
 }
