+++ src/main/java/com/harshraj9988/payment_simulator/entity/Account.java
@@ -15,5 +15,5 @@ public class Account {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     private String name;
-    private Double balance;
+    private Long balance;
 }
